pyproject.toml
[tool.poetry]
name = "yamli"
version = "0.1.0"
description = ""
authors = ["Louis Labeyrie <labeyrielouis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
streamlit = "^1.39.0"
pyyaml = "^6.0.2"
anytree = "^2.12.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yamli = "yamli.cli:main"
